Update db mapping.

+++ src/main.rs
@@ -24,7 +24,6 @@ use sp_core::hexdisplay::AsBytesRef;
 use crate::models::TxDirection::{AvailEth, EthAvail};
 use alloy::core::sol;
 use alloy::rpc::types::TransactionReceipt;
-use bigdecimal::BigDecimal;
 use http::Method;
 use jsonrpsee::{
     core::ClientError,
@@ -154,11 +153,11 @@ async fn transaction(
 
     let decoded = AvailBridgeEvents::decode_log(&log)?;
     let AvailBridgeEvents::MessageSent(call) = decoded.data;
-    let message_id: u64 = call.messageId.try_into()?;
+    let message_id: i64 = call.messageId.try_into()?;
 
     sqlx::query_file!(
         "sql/insert_tx.sql",
-        BigDecimal::from(message_id),
+        message_id,
         "MessageSent",
         BridgeStatusEnum::Initiated as BridgeStatusEnum, // cast if needed
         tx.from,
